Various fix
Former-commit-id: 7f86b51b1671bc7e7d635029a27f6300dd8c2186 [formerly aafb13aed0afea4e637f772c1d68013f18e2874e] [formerly a2adeeb59c5a3f287fe3a1db24ec6bca5fca88a1 [formerly e551e829edd399fa59160603ffdd608eec8b7e18] [formerly 656b85e6b3ca9871461e07a21bc5ce26ce73aec8] [formerly 4a2f1436ff4ec18bfff94622554e961abbf13a1e [formerly 82ce2c3959e1e07097317437945a5b2613b97303]] [formerly a2adeeb59c5a3f287fe3a1db24ec6bca5fca88a1 [formerly e551e829edd399fa59160603ffdd608eec8b7e18] [formerly 656b85e6b3ca9871461e07a21bc5ce26ce73aec8] [formerly 4a2f1436ff4ec18bfff94622554e961abbf13a1e [formerly 82ce2c3959e1e07097317437945a5b2613b97303]] [formerly 381475b13704b5387e66ddc663759b319bdaf914 [formerly 1ab423083703c983cd3a024ace3a25242b03a4ff]]]]
Former-commit-id: 715ca5d46cf7ea172bf4a43814fcf85ab6c27532
Former-commit-id: e4bdbfce60328288c4320972c58ec546bedbc034
Former-commit-id: 8e8eb62ab28237ac9e798c6584033ecf514f0d2a
Former-commit-id: c5a2ca0dbcf8f73cca050c80e2366d0f9f1cc6d9
Former-commit-id: 6673d31e8c199f0fd6a23ea4b6e9985692dd1708 [formerly a8696763bf9a84226c266ffcd90c2f851b9f102c]
Former-commit-id: 16f8cf32dc45b62721f75873448fcf39d44faeab
Former-commit-id: 68c8ed3dfe3864ab7e498a800022390430808450

+++ Source/Raw.Net/Source/Format/Tiff/IFD/Tag.cs
@@ -186,7 +186,7 @@ namespace RawNet
         {
             array = new float[num];
             for (int i = 0; i < num; i++)
-                array[i] = (float)Convert.ToDecimal(data[i]);
+                array[i] = Convert.ToSingle(data[i]);
         }
 
         public int getInt(int v)
